기본화면(하단바,상단로그)
