--- Memento-Project/lib/javafx.properties
@@ -1,3 +1,0 @@
-javafx.version=23
-javafx.runtime.version=23+29
-javafx.runtime.build=29
