--- Memento-Project/src/App.java
@@ -1,5 +1,0 @@
-public class App {
-    public static void main(String[] args) throws Exception {
-        System.out.println("Hello, World!");
-    }
-}
