+++ MementoProject/.vscode/launch.json
@@ -1,0 +1,15 @@
+{
+  "version": "0.2.0",
+  "configurations": [
+    {
+      "type": "java",
+      "name": "Launch Main",
+      "request": "launch",
+      "mainClass": "Main",
+      "vmArgs": "--module-path \"C:\\Program Files\\Java\\javafx-sdk-23\\lib\" --add-modules javafx.controls,javafx.fxml --enable-preview",
+      "classPaths": [
+        "C:\\Users\\yueun\\JavaProject\\Memento-Project\\MementoProject\\bin"
+      ]
+    }
+  ]
+}
