+++ MementoProject/.vscode/settings.json
@@ -1,7 +1,16 @@
 {
-    "java.project.sourcePaths": ["src"],
+    "java.project.sourcePaths": [
+        "src"
+    ],
     "java.project.outputPath": "bin",
     "java.project.referencedLibraries": [
-        "lib/**/*.jar"
+        "c:\\Users\\yueun\\Downloads\\openjfx-23_windows-x64_bin-sdk\\javafx-sdk-23\\lib\\javafx.web.jar",
+        "c:\\Users\\yueun\\Downloads\\openjfx-23_windows-x64_bin-sdk\\javafx-sdk-23\\lib\\javafx.base.jar",
+        "c:\\Users\\yueun\\Downloads\\openjfx-23_windows-x64_bin-sdk\\javafx-sdk-23\\lib\\javafx.controls.jar",
+        "c:\\Users\\yueun\\Downloads\\openjfx-23_windows-x64_bin-sdk\\javafx-sdk-23\\lib\\javafx.fxml.jar",
+        "c:\\Users\\yueun\\Downloads\\openjfx-23_windows-x64_bin-sdk\\javafx-sdk-23\\lib\\javafx.graphics.jar",
+        "c:\\Users\\yueun\\Downloads\\openjfx-23_windows-x64_bin-sdk\\javafx-sdk-23\\lib\\javafx.media.jar",
+        "c:\\Users\\yueun\\Downloads\\openjfx-23_windows-x64_bin-sdk\\javafx-sdk-23\\lib\\javafx.swing.jar",
+        "c:\\Users\\yueun\\Downloads\\openjfx-23_windows-x64_bin-sdk\\javafx-sdk-23\\lib\\javafx-swt.jar"
     ]
 }
