+++ MementoProject/bin/styles/styles.css
@@ -1,0 +1,34 @@
+/* styles.css */
+
+/* 전체 스타일 설정 */
+body {
+    -fx-font-family: "Segoe UI", sans-serif; /* 기본 폰트 설정 */
+}
+
+/* HBox에 적용되는 스타일 */
+hbox {
+    -fx-padding: 10; /* 내부 여백 */
+}
+
+/* 텍스트 스타일 */
+.text {
+    -fx-padding: 15; /* 텍스트 주변 여백 */
+    -fx-font-size: 20px; /* 폰트 크기 */
+    -fx-text-fill: white; /* 글자 색상 */
+    -fx-alignment: light; /* 중앙 정렬 */
+    -fx-spacing: 20px; /* 글자 간격 */
+    -fx-background-color: transparent; /* 배경색 투명 */
+}
+
+/* 하단 내비게이션 바 스타일 */
+.hbox {
+    -fx-background-color: #ff95ea; /* 하단 바 배경색 */
+    -fx-padding: 10; /* 하단 바 내부 여백 */
+    -fx-border-radius: 5; /* 모서리 둥글기 */
+}
+
+/* 버튼에 마우스 오버 효과 */
+.text:hover {
+    -fx-text-fill: #ffeb3b; /* 마우스 오버 시 글자 색상 변경 */
+    -fx-background-color: rgba(255, 255, 255, 0.2); /* 마우스 오버 시 배경색 변경 */
+}
