--- README.md
@@ -1,2 +1,0 @@
-# Memento-Project
-자바로 메멘토 다이어리 만들기 °˖✧◝(⁰▿⁰)◜✧˖°
